git-svn-id: https://svn.mothershipdev.com:9443/svn/PaymentSense_MothershipV2_Core/branches/RepositoryRefactoring@604 0f896bb5-3ab0-2d4e-82a7-85ef10020915

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Event/EventTransactionDto.cs
@@ -11,7 +11,7 @@ namespace PS.Mothership.Core.Common.Dto.Event
         [DataMember]
         public Guid ParentEventGuid { get; set; }
         [DataMember]
-        public int EventTypeKey { get; set; }
+        public long EventTypeKey { get; set; }
         [DataMember]
         public Guid StatusOwnerGuid { get; set; }
         [DataMember]
